chore: move repos code to repos file
